backend/requirements.txt
supabase
python-dotenv
fastapi
uvicorn
pose-format
spoken-to-signed
tensorflow
numpy
pandas
